+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/ui/guides/AddingCustomAttachments.kt
@@ -10,9 +10,10 @@ import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import com.google.android.material.datepicker.MaterialDatePicker
 import io.getstream.chat.android.models.Attachment
-import io.getstream.chat.android.models.ChannelCapabilities
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.ui.ChatUI
+import io.getstream.chat.android.ui.common.feature.messages.composer.capabilities.canSendMessage
+import io.getstream.chat.android.ui.common.feature.messages.composer.capabilities.canUploadFile
 import io.getstream.chat.android.ui.common.state.messages.Edit
 import io.getstream.chat.android.ui.common.state.messages.composer.MessageComposerState
 import io.getstream.chat.android.ui.feature.messages.composer.MessageComposerContext
@@ -85,8 +86,8 @@ class AddingCustomAttachments {
             }
 
             override fun renderState(state: MessageComposerState) {
-                val canSendMessage = state.ownCapabilities.contains(ChannelCapabilities.SEND_MESSAGE)
-                val canUploadFile = state.ownCapabilities.contains(ChannelCapabilities.UPLOAD_FILE)
+                val canSendMessage = state.canSendMessage()
+                val canUploadFile = state.canUploadFile()
                 val hasTextInput = state.inputValue.isNotEmpty()
                 val hasAttachments = state.attachments.isNotEmpty()
                 val hasCommandInput = state.inputValue.startsWith("/")